+++ test/realtime/integration/proxy/connection_open_failures_test.dart
@@ -66,8 +66,8 @@ void main() {
 
       final apiKey = testApp.keys[0].keyStr;
 
-      final client = PubSubClient(
-        options: ClientOptions(
+      final client = createClient(
+        ClientOptions(
           authCallback: makeAuthCallback(apiKey),
           endpoint: 'localhost',
           port: session.proxyPort,
@@ -127,8 +127,8 @@ void main() {
       final apiKey = testApp.keys[0].keyStr;
       var authCallbackCount = 0;
 
-      final client = PubSubClient(
-        options: ClientOptions(
+      final client = createClient(
+        ClientOptions(
           authCallback: makeAuthCallback(
             apiKey,
             onCalled: () => authCallbackCount++,
@@ -180,8 +180,8 @@ void main() {
       final apiKey = testApp.keys[0].keyStr;
       final stateChanges = <ConnectionStateChange>[];
 
-      final client = PubSubClient(
-        options: ClientOptions(
+      final client = createClient(
+        ClientOptions(
           authCallback: makeAuthCallback(apiKey),
           endpoint: 'localhost',
           port: session.proxyPort,
@@ -251,8 +251,8 @@ void main() {
 
       final apiKey = testApp.keys[0].keyStr;
 
-      final client = PubSubClient(
-        options: ClientOptions(
+      final client = createClient(
+        ClientOptions(
           authCallback: makeAuthCallback(apiKey),
           endpoint: 'localhost',
           port: session.proxyPort,
@@ -298,8 +298,8 @@ void main() {
 
       final apiKey = testApp.keys[0].keyStr;
 
-      final client = PubSubClient(
-        options: ClientOptions(
+      final client = createClient(
+        ClientOptions(
           authCallback: makeAuthCallback(apiKey),
           endpoint: 'localhost',
           port: session.proxyPort,