+++ test/realtime/integration/proxy/presence_reentry_test.dart
@@ -40,8 +40,8 @@ void main() {
         final session = await ProxySession.create();
         addTearDown(() async => await session.close());
 
-        final client = PubSubClient(
-          options: ClientOptions(
+        final client = createClient(
+          ClientOptions(
             authCallback: (params) async {
               return JwtHelper.generateToken(
                 apiKey: apiKey,
@@ -209,8 +209,8 @@ void main() {
       );
       addTearDown(() async => await session.close());
 
-      final client = PubSubClient(
-        options: ClientOptions(
+      final client = createClient(
+        ClientOptions(
           authCallback: (params) async {
             return JwtHelper.generateToken(
               apiKey: apiKey,